fix(auth_microservice) moved feign model to openfeign and added auth validation
