+++ auth/src/main/java/com/epam/esm/utils/EntityToDtoMapper.java
@@ -1,6 +1,6 @@
 package com.epam.esm.utils;
 
-import com.epam.esm.auth.models.CreateUserRequest;
+import com.epam.esm.utils.openfeign.CreateUserRequest;
 import com.epam.esm.auth.models.RegisterRequest;
 import org.mapstruct.Mapper;
